requirements.txt
PyYAML==3.10
docker-py==0.4.0
vcsinfo>=0.1.13

